refactor(M6 PR2): extract the worktree picker into controller/picker.rs
Move open_worktree_picker, handle_picker_intent, the agent-overlay hint
(herdr_overlay), and the PickerView projection into a submodule. re_root + the
async status refresh stay in mod.rs as core orchestration (the picker only
triggers them). Behavior-preserving. mod.rs ~2578.
